SCRIPT: Deflate v0.5.0—Rework to a pull-streaming, iterative model
This script represents a shift from the conventional Rebol approach to implementing format decoders. At its core is an iterative decoder designed to read Deflate streams in chunks of arbitrary size and on-demand.

As expected, a decoder of this nature will be comparatively slower in an interpreted environment compared to the built-in C decoder. However, this decoder's ability to read streams iteratively and maintain state between calls allows access to initial portions of content quickly and efficiently, and can approximate non-blocking behavior. In addition, it serves as a model for the interface that built-in C decoders could adopt in the future.